Exclude entities on demand

+++ de.tgmz.discogs/src/main/java/de/tgmz/discogs/load/ReleaseContentHandler.java
@@ -350,16 +350,20 @@ public class ReleaseContentHandler extends DiscogsContentHandler {
 		try {
 			super.save(o);
 		} catch (PersistenceException e) {
-			// Existing release with subTracks yields "Duplicate row was found and `ASSERT` was specified"
-			// so we must remove it first
-			LOG.warn("Unable to merge release {}, removing and inserting it", r.getId());
+			if ("Duplicate row was found and `ASSERT` was specified".equals(e.getMessage())) {
+				// Existing release with subTracks yields "Duplicate row was found and `ASSERT` was specified"
+				// so we must remove it first
+				LOG.warn("Unable to merge release {}, removing and inserting it", r.getId());
 			
-			DatabaseService.getInstance().inTransaction(em -> {
-				Release r0 = em.find(Release.class, ((Discogs) o).getId());
-				em.remove(r0);
-			});
-			
-			super.save(o);
+				DatabaseService.getInstance().inTransaction(em -> {
+					Release r0 = em.find(Release.class, ((Discogs) o).getId());
+					em.remove(r0);
+				});
+				
+				super.save(o);
+			} else {
+				LOG.error("Unable to merge release {}", r.getId(), e);
+			}
 		}
 	}
 }
